+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Donation link
 
+### Changed
+
+- Allow to choose completion date #83
+
+### Fixed
+
+- Don't allow due dates in the past and completion dates in the future
+
 ## [0.13.1] - 2025-06-10
 
 ### Changed